Move OpenStack extractors into its own directory
