+++ caso/extract/openstack/nova.py
