+++ setup.cfg
@@ -48,7 +48,7 @@ console_scripts =
     caso-extract = caso.cmd.extract:main
 
 caso.extractors =
-    nova = caso.extract.nova:OpenStackExtractor
+    nova = caso.extract.openstack.nova:OpenStackExtractor
 
 caso.messenger = 
     noop = caso.messenger.noop:NoopMessenger
